+++ rms/modules/__init__.py
@@ -1,0 +1,1 @@
+"""RMS Modules - Encoder and Qualia Engine."""